修复SCP卡输出 合并人物卡管理
